backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pdfplumber==0.11.4
python-multipart==0.0.12
aiofiles==24.1.0
openpyxl==3.1.5
python-jose[cryptography]==3.3.0
httpx==0.27.2
